refactor(project): migração completa para clean architecture - removida estrutura antiga (handlers, services, repositorios diretos, utils soltos) - adicionada camada domain/usecase/interface/infra seguindo padrão clean - atualizado docker-compose, migrations e docs
