--- .env.example
@@ -1,32 +1,0 @@
-# ============= App =============
-ENVIRONMENT=production
-PORT=8080
-CORS_ORIGIN=http://localhost:8080 # modificar dominio em produção
-
-# ============= DB (MySQL 8) =============
-DB_HOST=127.0.0.1
-DB_PORT=3306
-DB_USER=user
-DB_PASS=userpassword
-DB_NAME=mydatabase 
-
-# ============= JWT =============
-JWT_SECRET=troque-esta-chave 
-RT_SECRET=troque-esta-chave-refresh 
-ACCESS_TTL=24h
-REFRESH_TTL=168h # 7 dias
-
-# ============= LDAP (Desenvolvimento - simulando AD no docker) =============
-# LDAP_SERVER=ldap://127.0.0.1:389
-# LDAP_BASE=dc=rede,dc=sp
-# LDAP_USER=cn=admin,dc=rede,dc=sp
-# LDAP_PASS=Prodam0
-# LDAP_LOGIN_ATTR=uid
-
-# =============  Active Directory (Produção) =============
-LDAP_SERVER=ldap://10.10.65.242
-LDAP_DOMAIN=@rede.sp
-LDAP_BASE=DC=rede,DC=sp
-LDAP_USER=
-LDAP_PASS=
-LDAP_LOGIN_ATTR=sAMAccountName
