--- ldap-init/usuario1.ldif
@@ -1,11 +1,0 @@
-dn: ou=users,dc=rede,dc=sp
-objectClass: organizationalUnit
-ou: users
-
-dn: uid=usuario1,ou=users,dc=rede,dc=sp
-objectClass: inetOrgPerson
-cn: Usuario Um
-sn: Um
-uid: usuario1
-mail: usuario1@rede.sp
-userPassword: {SSHA}8oKr70ILDb74U2OOjtWNTruEKlGqWcxu #senha: senha123
